+++ viadot/tasks/__init__.py
@@ -25,6 +25,8 @@ from .sqlite import SQLiteInsert, SQLiteSQLtoDF, SQLiteQuery
 from .supermetrics import SupermetricsToCSV, SupermetricsToDF
 from .sharepoint import SharepointToDF
 from .cloud_for_customers import C4CReportToDF, C4CToDF
+from .prefect import GetFlowNewDateRange
+from .prefect_conformed import ReRunFailedFlow
 
 try:
     from .sap_rfc import SAPRFCToDF